requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.2.0
Pillow==10.0.0
requests
